[IOTDB-2172] Refactor for IoTDBSimpleQueryIT
