+++ integration/src/test/java/org/apache/iotdb/db/integration/IoTDBSimpleQueryIT.java
@@ -18,7 +18,6 @@
  */
 package org.apache.iotdb.db.integration;
 
-import org.apache.iotdb.db.conf.IoTDBDescriptor;
 import org.apache.iotdb.db.exception.metadata.MetadataException;
 import org.apache.iotdb.db.metadata.MManager;
 import org.apache.iotdb.db.metadata.mnode.IMeasurementMNode;
@@ -55,18 +54,14 @@ import static org.junit.Assert.fail;
 @Category({LocalStandaloneTest.class})
 public class IoTDBSimpleQueryIT {
 
-  boolean autoCreateSchemaEnabled;
-
   @Before
   public void setUp() throws Exception {
     EnvFactory.getEnv().initBeforeTest();
-    autoCreateSchemaEnabled = IoTDBDescriptor.getInstance().getConfig().isAutoCreateSchemaEnabled();
   }
 
   @After
   public void tearDown() throws Exception {
     EnvFactory.getEnv().cleanAfterTest();
-    IoTDBDescriptor.getInstance().getConfig().setAutoCreateSchemaEnabled(autoCreateSchemaEnabled);
   }
 
   @Test
@@ -930,33 +925,6 @@ public class IoTDBSimpleQueryIT {
         }
       }
     }
-  }
-
-  @Test
-  @Category({ClusterTest.class})
-  public void testPartialInsertionAllFailed() throws SQLException {
-
-    boolean autoCreateSchemaEnabled =
-        IoTDBDescriptor.getInstance().getConfig().isAutoCreateSchemaEnabled();
-    boolean enablePartialInsert = IoTDBDescriptor.getInstance().getConfig().isEnablePartialInsert();
-
-    try (Connection connection = EnvFactory.getEnv().getConnection();
-        Statement statement = connection.createStatement()) {
-      IoTDBDescriptor.getInstance().getConfig().setAutoCreateSchemaEnabled(false);
-      IoTDBDescriptor.getInstance().getConfig().setEnablePartialInsert(true);
-
-      statement.execute("SET STORAGE GROUP TO root.sg1");
-
-      try {
-        statement.execute("INSERT INTO root.sg1(timestamp, s0) VALUES (1, 1)");
-        fail();
-      } catch (IoTDBSQLException e) {
-        assertTrue(e.getMessage().contains("s0"));
-      }
-    }
-
-    IoTDBDescriptor.getInstance().getConfig().setEnablePartialInsert(enablePartialInsert);
-    IoTDBDescriptor.getInstance().getConfig().setAutoCreateSchemaEnabled(autoCreateSchemaEnabled);
   }
 
   @Test
